# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{groupId}/teachers/{teacherId}": {
      "delete": {
        "operationId": "GroupsController_unassignTeacher",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher unassigned"
          }
        },
        "summary": "Unassign a teacher from a group",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/subjects/{subjectId}/teachers/{teacherId}": {
      "delete": {
        "operationId": "SubjectsController_unassignTeacher",
        "parameters": [
          {
            "name": "subjectId",
            "required": true,
            "in": "path",
            "description": "Subject ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher unassigned"
          }
        },
        "summary": "Unassign a teacher from a subject",
        "tags": [
          "Subjects"
        ]
      }
    },
    "/teachers": {
      "post": {
        "description": "Creates a new teacher profile for an existing user. Only admins and center owners can create teacher profiles.",
        "operationId": "TeachersController_createTeacher",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Teacher profile data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teacher profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Teacher profile already exists for this user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new teacher profile",
        "tags": [
          "Teachers"
        ]
      },
      "get": {
        "description": "Retrieves a paginated list of all teachers. Only admins and center owners can access this endpoint.",
        "operationId": "TeachersController_getAllTeachers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by field",
            "schema": {
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name (partial match)",
            "schema": {
              "example": "John",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Teachers list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherListResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all teachers",
        "tags": [
          "Teachers"
        ]
      }
    },
    "/teachers/{id}": {
      "get": {
        "description": "Retrieves a teacher profile by its ID. Users can only access their own profile unless they are admins or center owners.",
        "operationId": "TeachersController_getTeacherById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Teacher profile ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get teacher profile by ID",
        "tags": [
          "Teachers"
        ]
      },
      "put": {
        "description": "Updates a teacher profile. Users can only update their own profile unless they are admins or center owners.",
        "operationId": "TeachersController_updateTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Teacher profile ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated teacher profile data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Teacher profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update teacher profile",
        "tags": [
          "Teachers"
        ]
      },
      "delete": {
        "description": "Deletes a teacher profile. Only admins and center owners can delete teacher profiles.",
        "operationId": "TeachersController_deleteTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Teacher profile ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "204": {
            "description": "Teacher profile deleted successfully"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete teacher profile",
        "tags": [
          "Teachers"
        ]
      }
    },
    "/teachers/user/{userId}": {
      "get": {
        "description": "Retrieves a teacher profile by user ID. Users can only access their own profile unless they are admins or center owners.",
        "operationId": "TeachersController_getTeacherByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get teacher profile by user ID",
        "tags": [
          "Teachers"
        ]
      }
    },
    "/teachers/{id}/views": {
      "post": {
        "description": "Increments the profile view counter for a teacher. This endpoint can be called by any authenticated user.",
        "operationId": "TeachersController_incrementProfileViews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Teacher profile ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-type",
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "CENTER"
              ],
              "default": "GLOBAL",
              "example": "GLOBAL"
            },
            "description": "Scope type (GLOBAL or CENTER)"
          },
          {
            "in": "header",
            "required": false,
            "name": "x-scope-id",
            "schema": {
              "type": "string",
              "example": "center-uuid"
            },
            "description": "Scope ID (center UUID, required if x-scope-type is CENTER)"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile views incremented successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Teacher not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Increment teacher profile views",
        "tags": [
          "Teachers"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS SaaS API",
    "description": "API documentation for the LMS SaaS platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID of the teacher",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "biography": {
            "type": "string",
            "description": "Teacher biography",
            "example": "Experienced mathematics teacher with 10+ years of teaching experience in advanced calculus and algebra."
          },
          "experienceYears": {
            "type": "number",
            "description": "Years of teaching experience",
            "example": 10,
            "minimum": 0,
            "maximum": 50
          },
          "specialization": {
            "type": "string",
            "description": "Teacher specialization or subject area",
            "example": "Mathematics, Advanced Calculus, Linear Algebra"
          }
        },
        "required": [
          "userId"
        ]
      },
      "TeacherResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Teacher profile ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "description": "User ID of the teacher",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "biography": {
            "type": "string",
            "description": "Teacher biography",
            "example": "Experienced mathematics teacher with 10+ years of teaching experience in advanced calculus and algebra."
          },
          "experienceYears": {
            "type": "number",
            "description": "Years of teaching experience",
            "example": 10
          },
          "specialization": {
            "type": "string",
            "description": "Teacher specialization or subject area",
            "example": "Mathematics, Advanced Calculus, Linear Algebra"
          },
          "profileViews": {
            "type": "number",
            "description": "Number of profile views",
            "example": 150
          },
          "rating": {
            "type": "number",
            "description": "Teacher rating (0-5)",
            "example": 4.5,
            "minimum": 0,
            "maximum": 5
          },
          "studentsCount": {
            "type": "number",
            "description": "Number of students taught",
            "example": 45
          },
          "centersCount": {
            "type": "number",
            "description": "Number of centers the teacher works at",
            "example": 2
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T10:30:00Z"
          },
          "user": {
            "type": "object",
            "description": "User information",
            "example": {
              "id": "123e4567-e89b-12d3-a456-426614174000",
              "name": "John Doe",
              "email": "john.doe@example.com"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "profileViews",
          "rating",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {
          "biography": {
            "type": "string",
            "description": "Teacher biography",
            "example": "Experienced mathematics teacher with 10+ years of teaching experience in advanced calculus and algebra."
          },
          "experienceYears": {
            "type": "number",
            "description": "Years of teaching experience",
            "example": 10,
            "minimum": 0,
            "maximum": 50
          },
          "specialization": {
            "type": "string",
            "description": "Teacher specialization or subject area",
            "example": "Mathematics, Advanced Calculus, Linear Algebra"
          }
        }
      },
      "TeacherListResponseDto": {
        "type": "object",
        "properties": {
          "teachers": {
            "description": "List of teachers",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeacherResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of teachers",
            "example": 25
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 10
          }
        },
        "required": [
          "teachers",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}